# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/topics/create": {
            "post": {
                "description": "Ushbu API orqali yangi topic yaratishingiz mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Create a new topic",
                "parameters": [
                    {
                        "description": "Create Topic request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topic.CreateTopicReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli yaratildi",
                        "schema": {
                            "$ref": "#/definitions/topic.CreateTopicResp"
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri ma'lumot kiritdingiz",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Serverda xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/topics/delete": {
            "delete": {
                "description": "Ushbu API orqali mavjud topicni o'chirishingiz mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Delete an existing topic",
                "parameters": [
                    {
                        "description": "Delete Topic request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topic.DeleteTopicReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mavzu muvaffaqiyatli o'chirildi",
                        "schema": {
                            "$ref": "#/definitions/topic.DeleteTopicResp"
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri ma'lumot kiritdingiz",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Serverda xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/topics/getAll": {
            "get": {
                "description": "Bu API barcha mavzularni qaytaradi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topics"
                ],
                "summary": "Get all topics",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1000,
                        "description": "Limit of topics (optional)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for topics (optional)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "description": "Filter for subjects",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topic.GetAllFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mavzular ro'yxati",
                        "schema": {
                            "$ref": "#/definitions/topic.GetAllTopicsResp"
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri ma'lumot kiritildi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Ichki xatolik",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/topics/update": {
            "put": {
                "description": "Ushbu API orqali mavjud topicni yangilashingiz mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Update an existing topic",
                "parameters": [
                    {
                        "description": "Update Topic request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topic.UpdateTopicReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli yangilandi",
                        "schema": {
                            "$ref": "#/definitions/topic.UpdateTopicResp"
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri ma'lumot kiritdingiz",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Serverda xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "topic.CreateTopicReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "topic.CreateTopicResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "topic.DeleteTopicReq": {
            "type": "object",
            "properties": {
                "topic_id": {
                    "type": "string"
                }
            }
        },
        "topic.DeleteTopicResp": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "topic.GetAllFilter": {
            "type": "object",
            "properties": {
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "topic.GetAllTopicsResp": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topic.Topic"
                    }
                }
            }
        },
        "topic.Topic": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "topic.UpdateTopicReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "topic.UpdateTopicResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Gateway",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}